+++ i2vgen-xl/main.py
@@ -47,13 +47,13 @@ def main(config_path):
     first_frame = frame_list[0]
 
     # Initialize pipeline
-    pipe = I2VGenXLPipeline.from_pretrained("ali-vilab/i2vgen-xl", torch_dtype=torch.float16, variant="fp16")
+    pipe = I2VGenXLPipeline.from_pretrained("ali-vilab/i2vgen-xl", torch_dtype=torch.float32, variant="fp16")
     
     # Handling multiple edited frames
     
     custom_unet = I2VGenXLUNet2(**pipe.unet.config)
     custom_unet.load_state_dict(pipe.unet.state_dict())
-    custom_unet = custom_unet.to(torch.float16)
+    custom_unet = custom_unet.to(torch.float32)
     pipe.unet = custom_unet
     
     pipe.to(device)
